pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aut-mvp"
version = "0.1.0"
description = "AI UI automation testing MVP"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "jinja2>=3.1.4",
  "pyyaml>=6.0.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["aut*"]